+++ sknetwork/classification/propagation.py
@@ -58,7 +58,7 @@ class Propagation(BaseClassifier):
     >>> labels_true = graph.labels
     >>> labels = {0: labels_true[0], 33: labels_true[33]}
     >>> labels_pred = propagation.fit_predict(adjacency, labels)
-    >>> np.round(np.mean(labels_pred == labels_true), 2)
+    >>> float(np.round(np.mean(labels_pred == labels_true), 2))
     0.94
 
     References
